requirements.txt
django
python-dotenv
dj-database-url